Clean up more compile time warnings.

+++ bhdrs/b.h
@@ -3,6 +3,8 @@
 #ifndef B_h_f3a79ec21e8ae24eef6ee57844d8f0cc
 #define B_h_f3a79ec21e8ae24eef6ee57844d8f0cc
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -62,12 +64,11 @@ typedef short intlet;
 
 #define IsSmallInt(v) (((int)(v)) & 1)
 #define SmallIntVal(v) (((int)(v) & ~1) / 2)
-#define MkSmallInt(i) ((value)((i)*2 | 1))
+#define MkSmallInt(i) ( (value) (((intptr_t) i)*2 | 1) )
 	/* (Can't use << and >> because their effect on negative numbers
 		is not defined.) */
-
 typedef struct value {HEADER; string *cts;} *value;
-
+								
 #define Hdrsize (sizeof(struct value)-sizeof(string))
 
 #define Type(v) (IsSmallInt(v) ? Num : (v)->type)
